Created a new class in Constants called BUNDLE_KEYS -It will store all bundle keys so i can reference them between classes.
Decided to use argument to paste in the data from the activity to fragment
Error and progress are now displayed in the MainActivity and not the fragment

The TopMovieFragment is now only created if we get a successful response from the server with the movie data.

Renamed a few methods  to be more descriptive.